move genMap and genFilter
