+++ src/zql/ivm/graph/operators/map-operator.ts
@@ -1,4 +1,5 @@
-import {Multiset, genMap} from '../../multiset.js';
+import {genMap} from '../../../util/iterables.js';
+import {Multiset} from '../../multiset.js';
 import {DifferenceStream} from '../difference-stream.js';
 import {LinearUnaryOperator} from './linear-unary-operator.js';
