+++ src/zql/ivm/multiset.ts
@@ -1,25 +1,3 @@
 export type Entry<T> = readonly [T, Multiplicity];
 export type Multiplicity = number;
 export type Multiset<T> = Iterable<Entry<T>>;
-
-export function genMap<T, U>(s: Iterable<T>, cb: (x: T) => U) {
-  return {
-    *[Symbol.iterator]() {
-      for (const x of s) {
-        yield cb(x);
-      }
-    },
-  };
-}
-
-export function genFilter<T>(s: Iterable<T>, cb: (x: T) => boolean) {
-  return {
-    *[Symbol.iterator]() {
-      for (const x of s) {
-        if (cb(x)) {
-          yield x;
-        }
-      }
-    },
-  };
-}
