+++ src/zql/util/iterables.ts
@@ -1,3 +1,25 @@
+export function genMap<T, U>(s: Iterable<T>, cb: (x: T) => U) {
+  return {
+    *[Symbol.iterator]() {
+      for (const x of s) {
+        yield cb(x);
+      }
+    },
+  };
+}
+
+export function genFilter<T>(s: Iterable<T>, cb: (x: T) => boolean) {
+  return {
+    *[Symbol.iterator]() {
+      for (const x of s) {
+        if (cb(x)) {
+          yield x;
+        }
+      }
+    },
+  };
+}
+
 export function* mapIter<T, U>(
   iter: Iterable<T>,
   f: (t: T, index: number) => U,
